# Optimize convert_UTF8_to_JSON for mostly ASCII strings