--- a/benchmark/encoder.rb
+++ b/benchmark/encoder.rb
@@ -17,8 +17,8 @@ def implementations(ruby_obj)
   state = JSON::State.new(JSON.dump_default_options)
 
   {
-    json: ["json", proc { JSON.dump(ruby_obj) }],
     json_state: ["json (reuse)", proc { state.generate(ruby_obj) }],
+    json: ["json", proc { JSON.dump(ruby_obj) }],
     oj: ["oj", proc { Oj.dump(ruby_obj) }],
     rapidjson: ["rapidjson", proc { RapidJSON.dump(ruby_obj) }],
   }
@@ -59,7 +59,8 @@ end
 benchmark_encoding "small nested array", [[1,2,3,4,5]]*10
 benchmark_encoding "small hash", { "username" => "jhawthorn", "id" => 123, "event" => "wrote json serializer" }
 
-# On these two benchmark we perform well.
+# On these three benchmarks we perform well. Either on par or very closely faster/slower
+benchmark_encoding "mixed utf8", ([("a" * 5000) + "€" + ("a" * 5000)] * 2000), except: %i(json_state)
 benchmark_encoding "twitter.json", JSON.load_file("#{__dir__}/data/twitter.json"), except: %i(json_state)
 benchmark_encoding "citm_catalog.json", JSON.load_file("#{__dir__}/data/citm_catalog.json"), except: %i(json_state)