--- a/ext/json/ext/generator/extconf.rb
+++ b/ext/json/ext/generator/extconf.rb
@@ -4,6 +4,7 @@ if RUBY_ENGINE == 'truffleruby'
   # The pure-Ruby generator is faster on TruffleRuby, so skip compiling the generator extension
   File.write('Makefile', dummy_makefile("").join)
 else
+  append_cflags("-std=c99")
   $defs << "-DJSON_GENERATOR"
   create_makefile 'json/ext/generator'
 end