--- a/ext/json/ext/generator/generator.c
+++ b/ext/json/ext/generator/generator.c
@@ -25,96 +25,77 @@ static ID i_to_s, i_to_json, i_new, i_pack, i_unpack, i_create_id, i_extend;
  * Everything else (should be UTF-8) is just passed through and
  * appended to the result.
  */
-static void convert_UTF8_to_JSON(FBuffer *out_buffer, VALUE in_string, bool out_script_safe)
+static void convert_UTF8_to_JSON(FBuffer *out_buffer, VALUE str, const bool escape_table[256], bool out_script_safe)
 {
     const char *hexdig = "0123456789abcdef";
     char scratch[12] = { '\\', 'u', 0, 0, 0, 0, '\\', 'u' };
 
-    const char *in_utf8_str = RSTRING_PTR(in_string);
-    unsigned long in_utf8_len = RSTRING_LEN(in_string);
+    const char *ptr = RSTRING_PTR(str);
+    unsigned long len = RSTRING_LEN(str);
 
     unsigned long beg = 0, pos;
 
-    for (pos =  0; pos < in_utf8_len;) {
-        uint32_t ch;
-        short ch_len;
-        bool should_escape;
-
-        /* UTF-8 decoding */
-        short i;
-        if      ((in_utf8_str[pos] & 0x80) == 0x00) { ch_len = 1; ch = in_utf8_str[pos];        } /* leading 1 bit is   0b0     */
-        else if ((in_utf8_str[pos] & 0xE0) == 0xC0) { ch_len = 2; ch = in_utf8_str[pos] & 0x1F; } /* leading 3 bits are 0b110   */
-        else if ((in_utf8_str[pos] & 0xF0) == 0xE0) { ch_len = 3; ch = in_utf8_str[pos] & 0x0F; } /* leading 4 bits are 0b1110  */
-        else if ((in_utf8_str[pos] & 0xF8) == 0xF0) { ch_len = 4; ch = in_utf8_str[pos] & 0x07; } /* leading 5 bits are 0b11110 */
-        else {
-            rb_raise(rb_path2class("JSON::GeneratorError"), "source sequence is illegal/malformed utf-8");
-        }
-
-        for (i = 1; i < ch_len; i++) {
-            ch = (ch<<6) | (in_utf8_str[pos+i] & 0x3F);
-        }
-
-        /* JSON policy */
-        should_escape =
-            (ch < 0x20) ||
-            (ch == '"') ||
-            (ch == '\\') ||
-            (out_script_safe && (ch == '/')) ||
-            (out_script_safe && (ch == 0x2028)) ||
-            (out_script_safe && (ch == 0x2029));
-
+    for (pos = 0; pos < len;) {
+        unsigned char ch = ptr[pos];
         /* JSON encoding */
-        if (should_escape) {
-            if (pos > beg) {
-                fbuffer_append(out_buffer, &in_utf8_str[beg], pos - beg);
-            }
-
-            beg = pos + ch_len;
+        if (escape_table[ch]) {
+#define FLUSH_POS(bytes) if (pos > beg) { fbuffer_append(out_buffer, &ptr[beg], pos - beg); } pos += bytes; beg = pos;
             switch (ch) {
-                case '"':  fbuffer_append(out_buffer, "\\\"", 2); break;
-                case '\\': fbuffer_append(out_buffer, "\\\\", 2); break;
-                case '/':  fbuffer_append(out_buffer, "\\/", 2); break;
-                case '\b': fbuffer_append(out_buffer, "\\b", 2); break;
-                case '\f': fbuffer_append(out_buffer, "\\f", 2); break;
-                case '\n': fbuffer_append(out_buffer, "\\n", 2); break;
-                case '\r': fbuffer_append(out_buffer, "\\r", 2); break;
-                case '\t': fbuffer_append(out_buffer, "\\t", 2); break;
-                default:
-                    if (ch <= 0xFFFF) {
+                case '"':  FLUSH_POS(1); fbuffer_append(out_buffer, "\\\"", 2); break;
+                case '\\': FLUSH_POS(1); fbuffer_append(out_buffer, "\\\\", 2); break;
+                case '/':  FLUSH_POS(1); fbuffer_append(out_buffer, "\\/", 2); break;
+                case '\b': FLUSH_POS(1); fbuffer_append(out_buffer, "\\b", 2); break;
+                case '\f': FLUSH_POS(1); fbuffer_append(out_buffer, "\\f", 2); break;
+                case '\n': FLUSH_POS(1); fbuffer_append(out_buffer, "\\n", 2); break;
+                case '\r': FLUSH_POS(1); fbuffer_append(out_buffer, "\\r", 2); break;
+                case '\t': FLUSH_POS(1); fbuffer_append(out_buffer, "\\t", 2); break;
+                default: {
+                    if ((ch & 0x80) == 0x00) { /* leading 1 bit is   0b0     */
+                        FLUSH_POS(1);
                         scratch[2] = hexdig[ch >> 12];
                         scratch[3] = hexdig[(ch >> 8) & 0xf];
                         scratch[4] = hexdig[(ch >> 4) & 0xf];
                         scratch[5] = hexdig[ch & 0xf];
                         fbuffer_append(out_buffer, scratch, 6);
+                    } else if ((ch & 0xE0) == 0xC0) { /* leading 3 bits are 0b110   */
+                        pos += 2;
+                    } else if ((ch & 0xF0) == 0xE0) { /* leading 4 bits are 0b1110  */
+                        unsigned char b2 = ptr[pos + 1];
+                        unsigned char b3 = ptr[pos + 2];
+                        if (out_script_safe && (b2 == 0x80)) {
+                            if (b3 == 0xA8) {
+                                FLUSH_POS(3);
+                                fprintf(stderr, "escape: \\u2028 pos = %ld\n", pos);
+                                fbuffer_append(out_buffer, "\\u2028", 6);
+                            } else if (b3 == 0xA9) {
+                                FLUSH_POS(3);
+                                fprintf(stderr, "escape: \\u2029 pos = %ld\n", pos);
+                                fbuffer_append(out_buffer, "\\u2029", 6);
+                            } else {
+                                pos += 3;
+                            }
+                        } else {
+                            pos += 3;
+                        }
+                    } else if ((ch & 0xF8) == 0xF0) { /* leading 5 bits are 0b11110 */
+                        pos += 4;
                     } else {
-                        uint16_t hi, lo;
-                        ch -= 0x10000;
-                        hi = 0xD800 + (uint16_t)(ch >> 10);
-                        lo = 0xDC00 + (uint16_t)(ch & 0x3FF);
-
-                        scratch[2] = hexdig[hi >> 12];
-                        scratch[3] = hexdig[(hi >> 8) & 0xf];
-                        scratch[4] = hexdig[(hi >> 4) & 0xf];
-                        scratch[5] = hexdig[hi & 0xf];
-
-                        scratch[8] = hexdig[lo >> 12];
-                        scratch[9] = hexdig[(lo >> 8) & 0xf];
-                        scratch[10] = hexdig[(lo >> 4) & 0xf];
-                        scratch[11] = hexdig[lo & 0xf];
-
-                        fbuffer_append(out_buffer, scratch, 12);
+                        // This should be unreachable
+                        rb_raise(rb_path2class("JSON::GeneratorError"), "source sequence is illegal/malformed utf-8");
                     }
+                }
             }
+        } else {
+            pos++;
         }
+    }
+#undef FLUSH_POS
 
-        pos += ch_len;
+    if (beg < len) {
+        fbuffer_append(out_buffer, &ptr[beg], len - beg);
     }
 
-    if (beg < in_utf8_len) {
-        fbuffer_append(out_buffer, &in_utf8_str[beg], in_utf8_len - beg);
-    }
-
-    RB_GC_GUARD(in_string);
+    RB_GC_GUARD(str);
 }
 
 static const bool escape_table[256] = {
@@ -736,7 +717,7 @@ static void generate_json_string(FBuffer *buffer, VALUE Vstate, JSON_Generator_S
             if (RB_UNLIKELY(state->ascii_only)) {
                 convert_UTF8_to_ASCII_only_JSON(buffer, obj, state->script_safe);
             } else {
-                convert_UTF8_to_JSON(buffer, obj, state->script_safe);
+                convert_UTF8_to_JSON(buffer, obj, state->script_safe ? script_safe_escape_table : escape_table, state->script_safe);
             }
             break;
         default: